Exit cleanly when killed externally
On Linux and Mac OS when a process is killed it receives a `SIGTERM`
signal (or `SIGINT` if force-killed). This change runs `process.exit()`
when receiving that signal, which causes the `handleExit` callback to
run, and causes child processes to exit.

Before this change when running on Linux killing the main process would
leave a child process running indefinitely.